move the gpu_stats up
