+++ lact-gui/src/app/gpu_stats_section.rs
