+++ lact-gui/src/app/pages/power_page.rs
@@ -3,10 +3,10 @@ mod power_cap_section;
 mod power_states;
 
 use super::PageUpdate;
-use crate::app::ext::{RelmDefaultLauchable, RelmLaunchable as _};
-use crate::app::pages::gpu_stats_section::{
+use crate::app::gpu_stats_section::{
     GpuStat, GpuStatsSection, GpuStatsSectionConfig, GpuStatsSectionMsg,
 };
+use crate::app::utils::ext::{RelmDefaultLauchable, RelmLaunchable as _};
 use adw::prelude::*;
 use amdgpu_sysfs::gpu_handle::{
     PerformanceLevel, PowerLevelKind, power_profile_mode::PowerProfileModesTable,
